requirements.txt
# Telegram Bot Framework
aiogram==3.15.0

# Environment variables
python-dotenv==1.0.1

# Database
aiosqlite==0.20.0

# Utilities
python-dateutil==2.9.0
